requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
python-dotenv==0.13.0
SQLAlchemy==1.3.17
flask_migrate
gunicorn
python-consul
flask_jwt_extended
flask_restx
jsons
psycopg2